+++ packages/frameworks/angular/src/combobox/combobox-trigger.directive.ts
@@ -4,8 +4,10 @@
 import {Component, computed} from "@angular/core"
 import {ChevronDown} from "lucide-angular"
 
+import {useInlineIconButtonApi} from "@qualcomm-ui/angular/inline-icon-button"
 import {CoreComboboxTriggerDirective} from "@qualcomm-ui/angular-core/combobox"
 import {provideIcons} from "@qualcomm-ui/angular-core/lucide"
+import {mergeProps} from "@qualcomm-ui/utils/merge-props"
 
 import {useQdsComboboxContext} from "./qds-combobox-context.service"
 
@@ -15,17 +17,32 @@ import {useQdsComboboxContext} from "./qds-combobox-context.service"
   standalone: false,
   template: `
     <ng-content>
-      <svg qIcon="ChevronDown"></svg>
+      <svg
+        qIcon="ChevronDown"
+        [q-bind]="inlineIconButtonApi().getIconBindings()"
+      ></svg>
     </ng-content>
   `,
 })
 export class ComboboxTriggerDirective extends CoreComboboxTriggerDirective {
   protected readonly qdsComboboxContext = useQdsComboboxContext()
+  protected readonly inlineIconButtonApi = useInlineIconButtonApi({
+    emphasis: "neutral",
+    size: computed(() =>
+      this.qdsComboboxContext().size === "sm" ? "sm" : "md",
+    ),
+    variant: "scale",
+  })
 
   constructor() {
     super()
     this.trackBindings.extendWith(
-      computed(() => this.qdsComboboxContext().getIndicatorBindings()),
+      computed(() =>
+        mergeProps(
+          this.qdsComboboxContext().getIndicatorBindings(),
+          this.inlineIconButtonApi().getRootBindings(),
+        ),
+      ),
     )
   }
 }