+++ packages/frameworks/angular/src/select/select-indicator.directive.ts
@@ -1,10 +1,12 @@
 // Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
 // SPDX-License-Identifier: BSD-3-Clause-Clear
 
-import {Component, computed, input} from "@angular/core"
+import {Component, computed, input, type OnInit} from "@angular/core"
 import {ChevronDown} from "lucide-angular"
 
+import {useInlineIconButtonApi} from "@qualcomm-ui/angular/inline-icon-button"
 import type {LucideIconOrString} from "@qualcomm-ui/angular-core/lucide"
+import {useTrackBindings} from "@qualcomm-ui/angular-core/machine"
 import {useSelectContext} from "@qualcomm-ui/angular-core/select"
 import {mergeProps} from "@qualcomm-ui/utils/merge-props"
 
@@ -14,10 +16,10 @@ import {useQdsSelectContext} from "./qds-select-context.service"
   selector: "[q-select-indicator]",
   standalone: false,
   template: `
-    <svg [q-bind]="indicatorBindings()" [qIcon]="icon()" />
+    <svg [q-bind]="inlineIconButtonApi().getIconBindings()" [qIcon]="icon()" />
   `,
 })
-export class SelectIndicatorDirective {
+export class SelectIndicatorDirective implements OnInit {
   /**
    * Indicator icon.
    *
@@ -27,11 +29,21 @@ export class SelectIndicatorDirective {
 
   protected readonly qdsSelectContext = useQdsSelectContext()
   protected readonly selectContext = useSelectContext()
+  protected readonly inlineIconButtonApi = useInlineIconButtonApi({
+    emphasis: "neutral",
+    size: computed(() => (this.qdsSelectContext().size === "sm" ? "sm" : "md")),
+    variant: "scale",
+  })
 
-  readonly indicatorBindings = computed(() =>
+  protected readonly trackBindings = useTrackBindings(() =>
     mergeProps(
       this.selectContext().getIndicatorBindings(),
       this.qdsSelectContext().getIndicatorBindings(),
+      this.inlineIconButtonApi().getRootBindings(),
     ),
   )
+
+  ngOnInit() {
+    this.trackBindings()
+  }
 }