+++ packages/common/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @qualcomm-ui/core
 
+## 1.0.3 (2025/11/13)
+
+### Bug Fixes
+
+- combobox input element focus on click
+- presence ssr, dialog ssr, collapsible ssr
+
 ## 1.0.2 (2025/11/12)
 
 ### Bug Fixes