+++ packages/frameworks/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @qualcomm-ui/react
 
+## 1.0.3 (2025/11/13)
+
+### Bug Fixes
+
+- add reset styles for elements and fix styles
+- use inline icon button scale variant for select indicator
+- use inline-icon-button scale variant for combobox trigger
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/qds-core, @qualcomm-ui/core]
+
 ## 1.0.2 (2025/11/12)
 
 ### Bug Fixes